% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_test_mitemap.R
\name{binom_test_mitemap}
\alias{binom_test_mitemap}
\title{Test binomial on MiteMap for presence in the half part of the odor}
\usage{
binom_test_mitemap(MiteMap, modality = "Modality", p.adjust_method = "BH")
}
\arguments{
\item{MiteMap}{(required) The result of import_mitemap ($resulting_data) for raw_data}

\item{modality}{The modality to separate test}

\item{p.adjust_method}{method for p-adjustement. See \code{\link[=p.adjust]{p.adjust()}}}
}
\value{
A tibble of results for binomial test
}
\description{
The test is run for each Modality values and each Farm with a p-adjustement
step.
}
\examples{
binom_test_mitemap(HH_example$resulting_data)
binom_test_mitemap(HH_example$resulting_data, p.adjust_method = "bonferroni")
}
\author{
Adrien Taudière
}
