% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_hull_mitemap.R
\name{convex_hull_mitemap}
\alias{convex_hull_mitemap}
\title{Map convex hull on MiteMap}
\usage{
convex_hull_mitemap(
  MiteMap,
  unity = 1,
  tbe = 3,
  plot_show = TRUE,
  probs_quantile = 0.68,
  min_nb_spatial_points = 2,
  each_point_count_one = FALSE,
  hull_col = "red",
  plot_center_of_mass = TRUE
)
}
\arguments{
\item{MiteMap}{(required) The result of import_mitemap ($resulting_data) for raw_data}

\item{unity}{(default = 1): size of square grid in mm}

\item{tbe}{(default = 3)}

\item{plot_show}{(Logical, default = TRUE) : Do we plot all mitemap ?}

\item{probs_quantile}{(default = 0.68)}

\item{min_nb_spatial_points}{: Minimum number of spatial point to keep the sample}

\item{each_point_count_one}{(Logical ; default = FALSE)}

\item{hull_col}{color of hull area}

\item{plot_center_of_mass}{(Logical, default = TRUE): Do we plot the center of mass?}
}
\value{
A dataframe
}
\description{
Map convex hull on MiteMap
}
\examples{
MM <- filter_mitemap(MM_example,
  center_x = -20, center_y = -20,
  bad_range_value_x = 50, bad_range_value_y = 50
)
ch <- convex_hull_mitemap(MM)
ch <- convex_hull_mitemap(MM, plot_show = FALSE)
full_join(ch, MM) |>
  ggplot(aes(x = Modality, y = hull_area)) +
  geom_boxplot()

full_join(ch, MM) |>
  group_by(File_name) |>
  summarise(
    hull_center_away_from_odor = sum(center_of_mass_x > 0) > 0,
    Modality = unique(Modality)
  ) |>
  ggplot(aes(fill = hull_center_away_from_odor, x = Modality)) +
  geom_bar()
}
\author{
Adrien Taudière
}
