% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_heatmap.R
\name{extract_heatmap}
\alias{extract_heatmap}
\title{Create folder(s) with heatmaps extracted from zip files optionally organized
by a factor}
\usage{
extract_heatmap(
  path_to_folder,
  factor = NULL,
  output_path = "Heatmap",
  verbose = TRUE,
  force = FALSE,
  heatmap_file_name = "_carte_thermique_",
  ...
)
}
\arguments{
\item{path_to_folder}{Path to folder}

\item{factor}{(character, default NULL) Set to a column name
to create subfolder for each levels of the modality. If left to NULL,
all heatmaps are extracted in the same folder.}

\item{output_path}{(default =  "Heatmap")}

\item{verbose}{(Logical, default = TRUE) If TRUE, the function print additional
information.}

\item{force}{Force overwriting the path to output_path}

\item{heatmap_file_name}{Name of the heatmap file in the zip file. Default is
"\emph{carte_thermique}" (sorry for french).}

\item{...}{Other params for be passed on to \code{\link[=import_mitemap]{import_mitemap()}}}
}
\value{
Nothing. Create files in folders.
}
\description{
Create folder(s) with heatmaps extracted from zip files optionally organized
by a factor
}
\examples{
\dontrun{
extract_heatmap(system.file("extdata", "mitemap_example", package = "MiteMapTools"),
  factor = "Treatment"
)
extract_heatmap(system.file("extdata", "mitemap_example", package = "MiteMapTools"),
  factor = "Biomol_sp", force = TRUE
)
extract_heatmap(system.file("extdata", "mitemap_example", package = "MiteMapTools"),
  clean = FALSE, force = TRUE
)
unlink("Heatmap", recursive = TRUE)
}
}
\author{
Adrien Taudière
}
