% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitemap_tools.R
\name{get_file_extension}
\alias{get_file_extension}
\title{Get the extension of a file}
\usage{
get_file_extension(file_path)
}
\arguments{
\item{file_path}{(required): path to a file}
}
\value{
The extension of a file.
}
\description{
Internally used in \code{\link[=rename_files_with_number]{rename_files_with_number()}} for ex.
Warning: don't work when there is '.' in the name of the file before the extension
}
\examples{
get_file_extension("my_file.csv")
get_file_extension("my.file.csv")
get_file_extension("my_file.csv.zip")
}
\author{
Adrien Taudière
}
