% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HH_example}
\alias{HH_example}
\title{example dataset from MiteMapTools for shape Half-half}
\format{
A list of 4 elements
\itemize{
\item \code{resulting_data}: a (possibly huge) tibble with metadata information and position in
x, y and time.
\item \code{files_not_in_csv}: a list of files not present in csv filenames from zip files.
\item \code{files_not_in_metada}: a list of files not funded in metadata.
\item \code{duplicate_file_name_in_metadata}: a list of duplicated file names in metadata.
}

The \code{resulting_data} slot is structured in 16 columns (The fifth ones are from zip files
and the other from metadata files). See \code{\link[=import_mitemap]{import_mitemap()}}.
\itemize{
\item File_name
\item The Date and hour
\item The name of the MiteMap (can be redundant with MiteMap_number)
\item Total time spent in the half containing the odor source (second)
\item Total time spent in the opposite half (second)
\item Time spent immobile in the half containing the odor source (second)
\item Time spent immobile in the opposite half (second)
\item Total distance traveled in the half containing the odor source (mm)
\item Total distance traveled in the opposite half (mm)
\item Boolean variable indicating if the individual has remained immobile since the last record (1 if immobile)
\item Run number
\item Date
\item Start time
\item Farm
\item MiteMap number
\item Bag
\item Modality
}
}
\source{
\url{https://onlinelibrary.wiley.com/doi/10.1002/jez.2651}
}
\usage{
data(HH_example)
}
\description{
Obtained using code :
\code{mm_HH <- suppressWarnings(import_mitemap(system.file("extdata", "POUL6", package = "MiteMapTools"), type_of_files = "HH"))}
}
\keyword{datasets}
