% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mitemap.R
\name{import_mitemap_from_multiple_folder}
\alias{import_mitemap_from_multiple_folder}
\title{Import mitemap from multiple folders}
\usage{
import_mitemap_from_multiple_folder(
  folders = NULL,
  type_of_files = NULL,
  path_to_metadata = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{folders}{A list of path}

\item{type_of_files}{A pattern to select the type of csv (either "raw_data", "CH" or "HH")
files using their name (e.g. "CH")}

\item{path_to_metadata}{A list of path}

\item{verbose}{(logical). If TRUE, print additional information.}

\item{...}{Other params for be passed on to \code{\link[=import_mitemap]{import_mitemap()}}}
}
\value{
A list of 4 elements. See ?\code{\link[=import_mitemap]{import_mitemap()}}
}
\description{
Import mitemap from multiple folders
}
\author{
Adrien Taudière
}
