% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap_folder.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Create folder(s) with heatmap}
\usage{
make_heatmap(
  path_to_folder,
  output_path = "Heatmap",
  path_to_metadata = NULL,
  format_metadata = "csv",
  modality = "Modality",
  subfolder_by_farm = TRUE,
  subfolder_by_modality = TRUE,
  dec = ",",
  skip = 1,
  colnames_metadata = c("Run_number", "File_name", "Date", "Start_time", "Farm",
    "MiteMap_number", "Bag", "Modality"),
  force = FALSE
)
}
\arguments{
\item{path_to_folder}{Path to folder}

\item{output_path}{(default =  "Heatmap")}

\item{path_to_metadata}{Path to metadata}

\item{format_metadata}{(either csv of xlsx)}

\item{modality}{(character, default "Modality") Set to a column names
to create subfolder for each levels of the modality}

\item{subfolder_by_farm}{(Logical, default = TRUE) create subfolder for
each farm}

\item{subfolder_by_modality}{(Logical, default = TRUE) create subfolder for
each modality set in parameter \code{modality}}

\item{dec}{decimal for the csv metadata files.}

\item{skip}{(default 1) skip the first line of the metadata files.}

\item{colnames_metadata}{Column names for metadata column. Note that File_name
is required and must correspond to the names of zip files.}

\item{force}{Force overwriting the path to output_path}
}
\value{
Nothing. Create files in folders.
}
\description{
Create folder(s) with heatmap
}
\examples{
make_heatmap(system.file("extdata", "POUL6", package = "MiteMapTools"))
unlink("Heatmap", recursive = TRUE)
}
\author{
Adrien Taudière
}
