% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_heatmap_folder.R
\name{extract_heatmap_from_multiple_folders}
\alias{extract_heatmap_from_multiple_folders}
\title{Create folders with heatmap from multiples folders}
\usage{
extract_heatmap_from_multiple_folders(folders, ...)
}
\arguments{
\item{folders}{A list of path}

\item{...}{Other params for be passed on to \code{\link[=extract_heatmap]{extract_heatmap()}}}
}
\value{
Nothing. Create files in folders.
}
\description{
Create folders with heatmap from multiples folders
}
\author{
Adrien Taudière
}
