% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MM_example}
\alias{MM_example}
\title{example dataset from MiteMapTools}
\format{
A list of 4 elements
\itemize{
\item \code{resulting_data}: a (possibly huge) tibble with metadata information and position in
x, y and time.
\item \code{files_not_in_csv}: a list of files not present in csv filenames from zip files.
\item \code{files_not_in_metada}: a list of files not funded in metadata.
\item \code{duplicate_file_name_in_metadata}: a list of duplicated file names in metadata.
}

The \code{resulting_data} slot is structured in 12 columns (The fifth ones are from zip files
and the other from metadata files). See \code{\link[=import_mitemap]{import_mitemap()}}.
\itemize{
\item File_name
\item The time in second (position is recorded every 0.2s)
\item The position in x (in mm)
\item The position in y (in mm)
\item Boolean variable indicating if the individual has remained immobile since the last record (1 if immobile)
\item Run number
\item Date
\item Start time
\item Farm
\item MiteMap number
\item Bag
\item Modality
}
}
\source{
\url{https://onlinelibrary.wiley.com/doi/10.1002/jez.2651}
}
\usage{
data(MM_example)
}
\description{
Obtained using code :
}
\details{
MM_example <- suppressWarnings(import_mitemap(system.file("extdata", "POUL6", package = "MiteMapTools")))

save(MM_example, file="data/MM_example.rda")
}
\keyword{datasets}
