% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vioplot_mitemap.R
\name{vioplot_mitemap}
\alias{vioplot_mitemap}
\title{Make violin plot of MiteMap data}
\usage{
vioplot_mitemap(MiteMap, modality = "Modality", wrap = NULL)
}
\arguments{
\item{MiteMap}{(required) The result of import_mitemap ($resulting_data) for raw_data}

\item{modality}{: A name of column present in the MiteMap to separate violin plot.}

\item{wrap}{: A name of column present in the MiteMap to wrap violin plot.}
}
\value{
A ggplot object
}
\description{
Make violin plot of MiteMap data
}
\examples{
MM_filtered_centered <- filter_mitemap(MM_example, center_x = -20, center_y = -20)
vioplot_mitemap(MM_filtered_centered)
vioplot_mitemap(MM_example$resulting_data, modality = "Farm", wrap = "Modality") +
  geom_boxplot(alpha = 0.1)
}
\author{
Adrien Taudière
}
