% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_hull_mitemap.R
\name{convex_hull_mitemap}
\alias{convex_hull_mitemap}
\title{Map convex hull on MiteMap}
\usage{
convex_hull_mitemap(
  MiteMap,
  unity = 1,
  tbe = 3,
  plot_show = TRUE,
  probs_quantile = 0.68,
  min_nb_spatial_points = 2,
  each_point_count_one = FALSE,
  hull_col = "red",
  plot_center_of_mass = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{MiteMap}{(required) The result of import_mitemap}

\item{unity}{(default = 1): size of square grid in mm}

\item{tbe}{(default = 3)}

\item{plot_show}{(Logical, default = TRUE) : Do we plot all mitemap ?}

\item{probs_quantile}{(default = 0.68)}

\item{min_nb_spatial_points}{: Minimum number of spatial point to keep the sample}

\item{each_point_count_one}{(Logical ; default = FALSE) If TRUE, each spatial point
count only for one time step instead. By default, if the mite stay 3t at a given
spatial location, the spatial point count 3 times in the construction of the convex
hull.}

\item{hull_col}{color of hull area}

\item{plot_center_of_mass}{(Logical, default = TRUE): Do we plot the center of mass?}

\item{verbose}{(Logical, default = TRUE) If TRUE, the function print additional}
}
\value{
A dataframe with convex hull information for each run (File_name)
and plot of convex hull for each run if plot_show = TRUE.
}
\description{
Map convex hull on MiteMap
}
\examples{
ch <- convex_hull_mitemap(MM_data)
ch <- convex_hull_mitemap(MM_data, plot_show = FALSE)
full_join(ch, MM_data) |>
  ggplot(aes(x = Treatment, y = hull_area)) +
  geom_boxplot()

full_join(ch, MM_data) |>
  group_by(File_name) |>
  summarise(
    hull_center_away_from_odor = sum(center_of_mass_x > 0) > 0,
    Treatment = unique(Treatment)
  ) |>
  ggplot(aes(fill = hull_center_away_from_odor, x = Treatment)) +
  geom_bar()
}
\author{
Adrien Taudière
}
