% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mitemap.R
\name{filter_mitemap}
\alias{filter_mitemap}
\title{Filter MiteMap result}
\usage{
filter_mitemap(
  MiteMap,
  HH = NULL,
  CH = NULL,
  first_seconds_to_delete = 2,
  bad_range_value_x = 50,
  bad_range_value_y = 50,
  max_x_value = 42,
  min_x_value = 0,
  max_y_value = 42,
  min_y_value = 0,
  maximum_time = 603,
  center_x = 0,
  center_y = 0
)
}
\arguments{
\item{MiteMap}{(required) The result of import_mitemap ($resulting_data) for raw_data}

\item{HH}{The result of \code{\link[=import_mitemap]{import_mitemap()}} ($resulting_data) for type_of_files = "HH". It remove run with (i) time in <0, time out <0 and time_in/(time_in + time_out) > 1 in the HH files.}

\item{CH}{The result of \code{\link[=import_mitemap]{import_mitemap()}} ($resulting_data) for type_of_files = "CH". It remove run with (i) time in <0, time out <0 and time_in/(time_in + time_out) > 1 in the CH files.}

\item{first_seconds_to_delete}{(default: 2) How many seconds we need to delete?}

\item{bad_range_value_x}{(default: 42) The range value of x that is superior to what we expect.
This param filter values on the File_name basis.}

\item{bad_range_value_y}{(default: 42) The range value of y that is superior to what we expect.
This param filter values on the File_name basis.}

\item{max_x_value}{(default: 42) Maximum value for x axis.
This param filter values on the point (one line) basis.}

\item{min_x_value}{(default: 0) Minimum value for x axis.
This param filter values on the point (one line) basis.}

\item{max_y_value}{(default: 42) Maximum value for y axis.
This param filter values on the point (one line) basis.}

\item{min_y_value}{(default: 0) Minimum value for y axis.
This param filter values on the point (one line) basis.}

\item{maximum_time}{(default: 603) Maximum time for one run (in seconds)).}

\item{center_x}{(int, default 0) Center the value of x by additioning center_x mm
to x.mm.}

\item{center_y}{(int, default 0) Center the value of y by additioning center_y mm
to y.mm.}
}
\value{
a filtered \verb{resulting table} of class tibble
}
\description{
Filter MiteMap result
}
\examples{
MM_filtered_centered <- filter_mitemap(MM_example,
  center_x = -20, center_y = -20,
  bad_range_value_x = 50, bad_range_value_y = 50
)
filter_mitemap(MM_example,
  HH = HH_example$resulting_data,
  bad_range_value_x = 50, bad_range_value_y = 50
)
}
