% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MM_data}
\alias{MM_data}
\title{example dataset from MiteMapTools}
\format{
A tibble.
}
\source{
\url{https://onlinelibrary.wiley.com/doi/10.1002/jez.2651}
}
\usage{
data(MM_data)
}
\description{
Obtained using code :
MM_data <- import_mitemap(
system.file("extdata", "mitemap_example_large", package = "MiteMapTools"),
file_name_column = "File (mite ID)", verbose = FALSE
)
save(MM_data, file="data/MM_data.rda")
}
\keyword{datasets}
