% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitemap_tools.R
\name{plot_ind_mitemap}
\alias{plot_ind_mitemap}
\title{Plot individual map}
\usage{
plot_ind_mitemap(
  MiteMap,
  ind_index = c(1),
  time_animation = FALSE,
  add_base_circle = FALSE,
  center_base_circle = c(0, 0),
  breaks_animation = 10,
  diameter_base_circle = 40,
  npoints_base_circle = 100,
  linewidth = 1.2,
  alpha_shadow = 0.4
)
}
\arguments{
\item{MiteMap}{MiteMap (required) The result of import_mitemap
($resulting_data) for raw_data}

\item{ind_index}{(required, vector of int, default = c(1))
The index numbers of the file(s) (individual(s)) you want to plot}

\item{time_animation}{(logical, default FALSE) Does the plot is animate
using gganimate ?}

\item{add_base_circle}{(logical, default FALSE) Does the base circle is plot ?}

\item{center_base_circle}{(int) The center of the base circle. Only used if
\code{center_base_circle} is TRUE.}

\item{breaks_animation}{(int) The number of breaks in animation. Only used if
\code{time_animation} is TRUE.}

\item{diameter_base_circle}{(int) The diameter of the base circle.
Only used if \code{center_base_circle} is TRUE.}

\item{npoints_base_circle}{(int) The number of point to draw the base circle.
Only used if \code{center_base_circle} is TRUE.}

\item{linewidth}{(int) The linewidth of the path}

\item{alpha_shadow}{(int \link{0,1}) Alpha value (transparency) for shadow path.
Only used if time_animation` is TRUE.}
}
\value{
A list of ggplot object
}
\description{
Plot individual map
}
\examples{
MM <- filter_mitemap(MM_example,
  center_x = -28, center_y = -22, max_x_value = 52, max_y_value = 52
)

p_l <- plot_ind_mitemap(MM, ind_index = c(1, 2))
p_l[[1]] + p_l[[2]] & theme(legend.position = "none")

p_l_base_c <- plot_ind_mitemap(MM, add_base_circle = TRUE, linewidth = 1.7)
p_l_base_c[[1]] + scale_color_gradient(name = "Speed", trans = "log10", low = "cyan", high = "red")

p_l_anim <- plot_ind_mitemap(MM, time_animation = TRUE, breaks_animation = 15)
p_l_anim[[1]]

p_l <- plot_ind_mitemap(MM, ind_index = c(1:6))
(p_l[[1]] + p_l[[2]] + p_l[[3]]) /
  (p_l[[4]] + p_l[[5]] + p_l[[6]]) +
  plot_layout(guides = "collect") & scale_color_gradient(
  name = "Speed", trans = "log1p", low = "cyan", high = "red",
  limits = c(
    min(unlist(lapply(p_l, function(x) {
      min(x$data$dist_from_i_minus_1, na.rm = T)
    }))),
    max(unlist(lapply(p_l, function(x) {
      max(x$data$dist_from_i_minus_1, na.rm = T)
    })))
  )
)

(p_l[[1]] + p_l[[2]] + p_l[[3]]) /
  (p_l[[4]] + p_l[[5]] + p_l[[6]]) +
  plot_layout(guides = "collect") & scale_color_viridis_c(name = "Speed", trans = "log1p", limits = c(0, 2))
}
\author{
Adrien Taudière
}
