% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitemap_tools.R
\name{rename_files_with_number}
\alias{rename_files_with_number}
\title{Rename files with number inside a folder}
\usage{
rename_files_with_number(
  path_to_folder,
  remove_zip_files_with_parenthesis = TRUE
)
}
\arguments{
\item{path_to_folder}{(required) Path to folder}

\item{remove_zip_files_with_parenthesis}{(logical, default TRUE). Do we
remove zip_files with parenthesis in the name?}
}
\value{
Nothing. Only modify files in a folder
}
\description{
Rename files with number inside a folder
}
\author{
Adrien Taudière
}
