% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_mitemap.R
\name{summarize_mitemap}
\alias{summarize_mitemap}
\title{Summarize MiteMap data at the individual (File) level}
\usage{
summarize_mitemap(
  MiteMap,
  num_cols = c("distance_from_previous", "speed_mm_s", "distance_from_sources",
    "in_left_half_HH", "in_left_half_CH", "turning_angle_clockwise", "turning_angle",
    "turning_angle_odor_clockwise", "turning_angle_odor", "turning_angle_ratio_odor")
)
}
\arguments{
\item{MiteMap}{(required) The result of import_mitemap}

\item{num_cols}{A vector of names of numeric columns to summarize with mean, sd, min and max.}
}
\value{
A summary table (tibble) with one line per File_name and summary statistics.

New columns are created with the suffixes corresponding to the applied summary function:
\itemize{
\item For selected numeric columns num_cols summary statistics include mean,
standard deviation, minimum, and maximum values.
\item For character and factor columns, the first unique value is retained.
As file_name (individual mite) have unique metadata, all factor columns
have one value by file_name.
\item For logical columns, the proportion (mean) and number of TRUE values is
calculated. The number of FALSE is also calculated.
\item For other numeric columns, the mean value is calculated.
}
}
\description{
Summarize MiteMap data at the individual (File) level
}
\examples{
mm_csv <- suppressWarnings(import_mitemap(
  system.file("extdata", "mitemap_example", package = "MiteMapTools"),
  file_name_column="File (mite ID)"
))
sum_mm <- summarize_mitemap(mm_csv)
ggplot(sum_mm, aes(x=Treatment, y=distance_from_sources_mean, fill=Treatment)) +
ggrain::geom_rain() +
coord_flip()
}
\author{
Adrien Taudière
}
