% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vioplot_mitemap.R
\name{vioplot_mitemap}
\alias{vioplot_mitemap}
\title{Make violin plot of MiteMap data}
\usage{
vioplot_mitemap(MiteMap, factor = NULL, wrap = NULL, prop_points = NULL)
}
\arguments{
\item{MiteMap}{(required) The result of import_mitemap}

\item{factor}{: A name of column present in the MiteMap to separate violin plot.}

\item{wrap}{: A name of column present in the MiteMap to wrap violin plot.}

\item{prop_points}{(numeric between 0 and 1, default=NULL) If not NULL,
a proportion of points is randomly sampled and added to the violin plot.
This can help to visualize the data distribution, especially with large datasets.}
}
\value{
A ggplot object
}
\description{
Make violin plot of MiteMap data
}
\examples{
vioplot_mitemap(MM_data, "Treatment")
vioplot_mitemap(MM_data, factor = "Treatment", wrap = "Biomol_sp") +
  geom_boxplot(col = "gray60", width = 0.1)

vioplot_mitemap(MM_data, "Treatment", prop_points = 0.01)
}
\author{
Adrien Taudière
}
